+++ base/floatfuncs.jl
@@ -43,7 +43,7 @@ maxintfloat() = maxintfloat(Float64)
 isinteger(x::AbstractFloat) = (x - trunc(x) == 0)
 
 """
-    round([T,] x, [digits, [base]], [r::RoundingMode])
+    round([T,] x, [digits,] [r::RoundingMode;] base = 10)
 
 Rounds `x` to an integer value according to the provided
 [`RoundingMode`](@ref), returning a value of the same type as `x`. When not
@@ -71,14 +71,14 @@ rounded.
 [`InexactError`](@ref) if the value is not representable.
 
 `round(x, digits)` rounds to the specified number of digits after the decimal place (or
-before if negative). `round(x, digits, base)` rounds using a base other than 10.
+before if negative). `round(x, digits, base = base)` rounds using a base other than 10.
 
 # Examples
 ```jldoctest
 julia> round(pi, 2)
 3.14
 
-julia> round(pi, 3, 2)
+julia> round(pi, 3, base = 2)
 3.125
 ```
 
@@ -141,7 +141,7 @@ function _signif_og(x, digits, base)
 end
 
 """
-    signif(x, digits, [base])
+    signif(x, digits; base = 10)
 
 Rounds (in the sense of [`round`](@ref)) `x` so that there are `digits` significant digits, under a
 base `base` representation, default 10.
@@ -151,11 +151,11 @@ base `base` representation, default 10.
 julia> signif(123.456, 2)
 120.0
 
-julia> signif(357.913, 4, 2)
+julia> signif(357.913, 4, base = 2)
 352.0
 ```
 """
-function signif(x::Real, digits::Integer, base::Integer=10)
+function signif(x::Real, digits::Integer; base::Integer = 10)
     digits < 1 && throw(DomainError(digits, "`digits` cannot be less than 1."))
 
     x = float(x)
@@ -171,7 +171,7 @@ end
 
 for f in (:round, :ceil, :floor, :trunc)
     @eval begin
-        function ($f)(x::Real, digits::Integer, base::Integer=10)
+        function ($f)(x::Real, digits::Integer; base::Integer=10)
             x = float(x)
             og = convert(eltype(x),base)^digits
             r = ($f)(x * og) / og